requirements.txt
setuptools==69.2.0
scikit-learn==1.3.2
pandas==2.1.4
transformers==4.40.0
tokenizers==0.19.1
datasets==2.19.0
sentence-transformers==2.7.0
tensorboardX==2.6.1
accelerate==0.29.3
deepspeed==0.14.1
tqdm==4.66.2
pytest==7.4.3
scipy==1.12.0
seaborn==0.13.2
pillow>=10.3.0
numpy==1.26.3
matplotlib==3.8.2
packaging==24.0
wandb==0.16.3
sentencepiece==0.2.0
bitsandbytes==0.42.0
peft==0.10.0
opencv-python==4.9.0.80
flash-attn==2.5.6
hf_transfer==0.1.6